Classifier/requirements.txt
absl-py==1.4.0
asgiref==3.7.2
astunparse==1.6.3
backports.zoneinfo==0.2.1;python_version<"3.9"
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
Django==3.2.19
django-background-tasks==1.2.5
django-compat==1.0.15
flatbuffers==23.5.26
gast==0.4.0
google-auth==2.20.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.0
gunicorn==20.1.0
h5py==3.9.0
idna==3.4
importlib-metadata==6.7.0
jax==0.4.13
keras==2.12.0
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.3
ml-dtypes==0.2.0
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
Pillow==9.5.0
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
six==1.16.0
sqlparse==0.4.4
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
typing_extensions==4.6.3
urllib3==1.26.16
Werkzeug==2.3.6
wrapt==1.14.1
zipp==3.15.0